requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
python-dotenv==1.0.1
anthropic>=0.40.0
openai>=1.54.0
pytest==8.3.3
